+++ NEWS.md
@@ -4,6 +4,9 @@
 - `NoBMA` and `NoBMA.reg()` functions as wrappers around `RoBMA` `RoBMA.reg()` functions for simpler specification of publication bias unadjusted Bayesian model-averaged meta-analysis
 - adding odds ratios output transformation` 
 
+### Fixes
+- handling of non-converged models
+
 ## version 3.0.1
 ### Fixes (thanks to Don & Rens)
 - compilation issues with Clang (https://github.com/FBartos/RoBMA/issues/28)